src/requirements.txt
requests==2.31.0
pandas==2.2.1
pyarrow==15.0.2
google-cloud-storage==2.16.0
